flho-163 storage-add-another-address (#43)
* flho-163 storage-add-another-address. This step is to allow the user to add multiple addresses whilst using the same steps to loop over.
What is basically happening is, when we get to the address-lookup step or manual-address step, we are adding the address to an array along with an id and then clearing the fields in order to go around again.
When the user has added an address, the address will appear on the add-another-address step where they are able to change the address, add more or continue.
If the user chooses to add another address, they are taken to the storage-postcode step where they can do the process again. The user will be able to see the addresses the have added just like the add-another-address step.
If the user choose to Change the address, then they are taken to the storage-postcode step with the fields entered with the relevant data. The url is appended with /edit/${id}. The id allows us to get the relevant data to fill the fields and allows us to overwrite the old address with the new one in the array.
As well as this, the usual step components have been added:
	* The steps have been added to the routes config,
	* The fields have been added to the fields config,
	* Controllers have been added for the steps,
	* The translations have been added for the page, fields and validation,
	* Custom html templates have been created,
	* Acceptance tests have been updated for storage-address including the page-object.