Ensure translations are updated in dev mode when lang files change
The previous setup meant that the lang files were compiled when files changed, but did not trigger a restart of the server, so it would have to be restarted manually in order to see text changes. By including the lang file compilation in the main app start-up we can ensure that the running app in dev mode always shows the most recent content.

This does cause a slight overhead for restarts due to regular code changes, but it is minimal.