Mark address lookup postcode steps as complete immediately
Then throw missing prerequisites on select pages if no addresses are found to redirect back to lookup.

Because a user can click straight through to manual address entry the postcode input was not being set as complete. This meant that going back to edit the previous step would force a user to continue into the address lookup, rather than being returned to the confirm page.