Amend address errors on manual address step
The `*-address` field exists on both postcode lookup select and manual entry steps, but with different interactions in each case. This means the validation error messagesneed to be different.

As such, append "-manual" to the key for the address errors on manual steps so they can load a different set of error messages.