Add behaviour for submission to iCasework
Provides a behaviour that makes a POST to iCasework's create case API as part of `saveValues`. The behaviour allows for a custom `prepare` method to be passed as part of the behaviour configuration which can extend the default auth attributes with the appropriate case data for each of our applications.

```js
// in steps config
const CaseWorkSubmission = require('../common/behaviours/casework-submission');

'/confirm': {
  behaviours: [
    'complete',
    CaseWorkSubmission({
      prepare: (session) => { /* return data fields */ }
    })
  ],
  next: '/confirmation'
  ...
}
```