Fix storage address editing
The behaviour when editing storage addresses was a bit broken. It wasn't possible to add a new address when in an edit journey. Deleting an address broke the user out of the edit journey and left them on a standard journey.

Adds a "continueOnEdit" flag to the "yes" branch of the "add more addresses" step so that editing and selecting "yes" will allow the user to continue around that loop.

Swaps the delete functionality to being query strong based so that it does not remove the `/edit` parameter from the url.