Handle icasework failing with 200
In some circumstances icasework submission fails, but returns a 200 with an empty response. In these cases the line which accesses `data.createcaseresponse.caseid` fails because the JSON response is empty. By moving the call to next into a catch rather than a second argument we can ensure that these errors are caught and the request pipeline is not broken.