Implement email address checking in document upload app
Does a call to the `getcasedetails` API from iCasework for a reference number, and retrieves the email address associated with the case. Then adds a validation check to the email field which reports an error if the email address does not match.

Note: if the iCasework API returns an error then it is swallowed in order to prevent the enumeration of reference numbers or email addresses.