Allow users to cancel out of an upload journey
If a user has previously uploaded documents, and selects "Yes" to upload more documents then allow them to cancel and continue with their journey without adding more documents.

Do this by marking the upload step as complete on `get` if there are documents on the session. This allows the user to go directly to the next step without completing the form.