Update modules to latest versions
This was just going to be the theme, but npm@5 had been quietly saving some other updates, and it doesn't hurt to include those too.

Also remove package-lock.json because it's not being used for deploys, so managing it all the time is overhead. If we want to use it for deploys then we can un-ignore it in the future.