Throw timeout errors when calling icasework apis
Failures of IP whitelisting with icasework manifest with a long pause (request takes ~120s) followed by an nginx "bad gateway" error due to it hitting its default timeout.

By setting an explicit timeout on our calls to icasework we can send a timeout error page before nginx fails, and so prevent the nginx error page from displaying.

Additionally we prevent catching the error when calling `getcasedetails` and so allow the app to fail early - rather than allowing the user to continue and then displaying an incorrect credentials error later. Since the timeout error is not related to sending a good or bad reference this does not represent a security issue.